Refactored tests for Pet classes
